// src/lib/inspiration/common/enums.ts
export enum MoveKey {
  Up = 'ArrowUp',
  Left = 'ArrowLeft',
  Right = 'ArrowRight',
  Down = 'ArrowDown'
}
export enum MouseButton {
  Left = 0,
  Right = 2
}

/**
 * 素材类型
 */
export enum AssetType {
  'Image' = 'Image',
  'Json' = 'Json',
  'Graph' = 'Graph'
}


// 连接线 类型
export enum LinkType {
  'auto' = 'auto',
  'straight' = 'straight', // 直线
  'manual' = 'manual' // 手动折线
}


export enum ShutcutKey {
  Del = 'Delete',
  C = 'KeyC',
  V = 'KeyV',
  Z = 'KeyZ',
  A = 'KeyA',
  R = 'KeyR',
  Esc = 'Escape',
  Backspace = 'Backspace'
}